:dancers: Provide MySQL implementation
This commit adds MySQL Support to Tern. The implementation is based on
the Postgresql implementation in that it implements the Migrator
protocol.

Because MySQL is resoundingly terrible with Timestamps (second only
precision is basically useless), the entry in the schema migrations table
has been calculated in the code and inserted into the table by hand.

It would be nice to be able to use built in database functions for this
as you would in a more feature complete database engine.